Split file in module.
